+++ test-helpers/loadTests.js
@@ -4,6 +4,9 @@ import {jml} from '../src/jml.js';
 import jmlTests from '../test/test.jml.js';
 import otherMethodsTests from '../test/test.other-methods.js';
 import toJMLTests from '../test/test.toJML.js';
+import XMLSerializer from '../polyfills/XMLSerializer.js';
+
+window.XMLSerializer = XMLSerializer; // Needed by Firefox for consistency in serializing DOCTYPE (with undefined systemId/publicId)
 
 window.jml = jml;